+++ src/composables/index.ts
@@ -5,6 +5,7 @@
 /** Barrel re-exporting every Vue 3 composable. */
 export * from "./useAppLockState";
 export * from "./useLockState";
+export * from "./useNavDirection";
 export * from "./useOverlayBackHandler";
 export * from "./usePullToRefresh";
 export * from "./useSecretReveal";